--- src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/entity/MessageSender.java
@@ -1,5 +1,0 @@
-package bntu.fitr.gorbachev.ticketsgenerator.main.entity;
-
-public interface MessageSender {
-    void send(String msg);
-}
