+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/entity/impl/sender/MessageSender.java
@@ -1,0 +1,5 @@
+package bntu.fitr.gorbachev.ticketsgenerator.main.entity.impl.sender;
+
+public interface MessageSender {
+    void send(String msg);
+}
