+++ src/main/java/bntu/fitr/gorbachev/ticketsgenerator/main/entity/impl/sender/RegistrarSenderMessage.java
@@ -1,4 +1,4 @@
-package bntu.fitr.gorbachev.ticketsgenerator.main.entity;
+package bntu.fitr.gorbachev.ticketsgenerator.main.entity.impl.sender;
 
 public interface RegistrarSenderMessage {
     void add(MessageSender sender);
